requirements.txt
aiofiles==0.7.0
aiogram==3.0.0b1
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==5.0.0
charset-normalizer==2.0.11
frozenlist==1.3.0
idna==3.3
loguru==0.6.0
magic-filter==1.0.5
multidict==6.0.2
peewee==3.14.8
pyaes==1.6.1
pydantic==1.9.0
Pyrogram==1.4.0
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==2.0.3
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
TgCrypto==1.2.3
typing_extensions==4.0.1
yarl==1.7.2
lxml
fake-headers==1.0.2
asyncstdlib==3.10.5
pandas==1.3.2
XlsxWriter==3.0.1
APScheduler==3.9.1
cloudscraper==1.2.64
asyncstdlib==3.10.5
PyVirtualDisplay==3.0
undetected-chromedriver==2.1.1
asgiref